requirements.txt
# Core dependencies for JBL-Suno-DJ POC
sunoai>=1.0.7
rich>=13.7.0
click>=8.1.0
python-dotenv>=1.0.0
requests>=2.31.0

# Audio processing (for future integration)
sounddevice>=0.4.0
numpy>=1.26.0
ffmpeg-python>=0.2.0

# Speech recognition (for future integration)
whispercpp>=0.0.17
